move Premium to entity package
