+++ src/main/java/school/faang/user_service/entity/User.java
@@ -13,7 +13,7 @@ import school.faang.user_service.entity.event.Event;
 import school.faang.user_service.entity.goal.Goal;
 import school.faang.user_service.entity.goal.GoalInvitation;
 import school.faang.user_service.entity.event.Rating;
-import school.faang.user_service.service.premium.util.Premium;
+import school.faang.user_service.entity.premium.Premium;
 import school.faang.user_service.entity.promotion.UserPromotion;
 import school.faang.user_service.entity.recommendation.Recommendation;
