+++ src/main/java/school/faang/user_service/repository/premium/PremiumRepository.java
@@ -2,7 +2,7 @@ package school.faang.user_service.repository.premium;
 
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
-import school.faang.user_service.service.premium.util.Premium;
+import school.faang.user_service.entity.premium.Premium;
 
 import java.time.LocalDateTime;
 import java.util.List;
