+++ src/main/java/school/faang/user_service/service/premium/PremiumCheckService.java
@@ -7,7 +7,7 @@ import org.springframework.transaction.annotation.Transactional;
 import school.faang.user_service.dto.payment.PaymentResponse;
 import school.faang.user_service.entity.User;
 import school.faang.user_service.entity.payment.PaymentStatus;
-import school.faang.user_service.service.premium.util.Premium;
+import school.faang.user_service.entity.premium.Premium;
 import school.faang.user_service.entity.premium.PremiumPeriod;
 import school.faang.user_service.exception.payment.UnSuccessPaymentException;
 import school.faang.user_service.exception.premium.PremiumCheckFailureException;
