+++ src/main/java/school/faang/user_service/service/premium/PremiumService.java
@@ -9,7 +9,7 @@ import school.faang.user_service.entity.User;
 import school.faang.user_service.entity.premium.PremiumPeriod;
 import school.faang.user_service.repository.premium.PremiumRepository;
 import school.faang.user_service.service.payment.PaymentService;
-import school.faang.user_service.service.premium.util.Premium;
+import school.faang.user_service.entity.premium.Premium;
 import school.faang.user_service.service.premium.util.PremiumBuilder;
 
 @Slf4j
