+++ src/test/java/school/faang/user_service/service/premium/PremiumServiceTest.java
@@ -13,7 +13,7 @@ import school.faang.user_service.entity.payment.PaymentStatus;
 import school.faang.user_service.entity.premium.PremiumPeriod;
 import school.faang.user_service.repository.premium.PremiumRepository;
 import school.faang.user_service.service.payment.PaymentService;
-import school.faang.user_service.service.premium.util.Premium;
+import school.faang.user_service.entity.premium.Premium;
 import school.faang.user_service.service.premium.util.PremiumBuilder;
 
 import java.time.LocalDateTime;
